requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-http
apache-airflow==2.9.1
apache-airflow-providers-postgres
requests
psycopg2-binary
pendulum